fix(opportunities): correct type issues in calculators
Fix compilation errors:
- profit_taking: Handle float64 Quantity from Position
- rebalance_sells/buys: Remove Country field references (not in Security model)
- weight_based: Fix domain imports and Position type references

All modules now compile successfully.

+++ trader-go/internal/modules/opportunities/calculators/profit_taking.go
@@ -99,14 +99,14 @@ func (c *ProfitTakingCalculator) Calculate(
 		// Calculate quantity to sell
 		quantity := position.Quantity
 		if sellPercentage < 1.0 {
-			quantity = int(float64(position.Quantity) * sellPercentage)
+			quantity = position.Quantity * sellPercentage
 			if quantity == 0 {
 				quantity = 1
 			}
 		}
 
 		// Calculate value
-		valueEUR := float64(quantity) * currentPrice
+		valueEUR := quantity * currentPrice
 
 		// Apply transaction costs
 		transactionCost := ctx.TransactionCostFixed + (valueEUR * ctx.TransactionCostPercent)
@@ -138,7 +138,7 @@ func (c *ProfitTakingCalculator) Calculate(
 			Side:     "SELL",
 			Symbol:   position.Symbol,
 			Name:     security.Name,
-			Quantity: quantity,
+			Quantity: int(quantity),
 			Price:    currentPrice,
 			ValueEUR: netValueEUR,
 			Currency: string(security.Currency),
